Some more localization (now with formatting) Fixed bug in PDFObject where ByteBuffer.array isn't supported (the pre-5.0 version supports it, native version does not). FilePicker now handles the SeperatorField properly and is less likely to throw an exception on setting row height.
Main component is the update to PDFImage to the latest version. This has had a ample of of stuff changed, not a ton since some of the original code was still valid. Note that if any image is "illegal" and has components that are less then a byte and can cross byte bounds, it won't process correctly (right now).